feat(categories): add categories list
