+++ .gitignore
@@ -5,6 +5,7 @@
 /.pnp
 .pnp.js
 .env
+.idea
 
 # testing
 /coverage
